feat(config): update keybindings and theme
This commit introduces several updates to the Hyprland and Waybar configurations.

In Hyprland, the default browser is changed to Brave, and new keybindings are added for swapping windows and taking screenshots with different modes.

For Waybar, the theme has been updated by renaming the style and color files to partials, and the font family is updated to include Material Symbols. The module configuration is also streamlined, with an updated workspace format and new network icons.